Split oversized baseline gate test file
